apispec Convert video monitoring unit tests to integration tests.
